Added a background image
Placed a background image inside the body of the global CSS file. I set it to position: fixed and also added a dark layer on top of it to keep the readability of the website good.

I made the <Header> transparent and gave it a max-width. I want the header to not cover the entire screen, at least on bigger monitors. I slightly adjusted the dropdown menu's width to match the new styling choices I made.